Refactor bankid api code
- Refactor collect api logic to separate file
- Remove old, unused files
