+++ src/http/apis/bankid.ts
@@ -7,10 +7,9 @@ import {
 } from "../../bankid/types";
 import * as http from "../index";
 
-import { getCollectUrl } from "./bankid/utils";
-import { cancel as doCancelRequest } from "./bankid/cancel";
 import { auth as doAuthRequest } from "./bankid/auth";
-import { buildCollectRequest } from "./bankid/request-builder";
+import { collect as doCollectRequest } from "./bankid/collect";
+import { cancel as doCancelRequest } from "./bankid/cancel";
 
 const validate = <T>(response: http.IApiResponse<T>): T => {
     if (!response || !response.successful || !response.response) {
@@ -27,14 +26,8 @@ export const auth = (
 
 export const collect = (
     requestOptions: IBankIdCollectRequest
-): Promise<IBankIdCollectApiResponse> => {
-    const url = getCollectUrl(requestOptions.orderRef);
-    const request = buildCollectRequest();
-
-    return http
-        .captureStateContext(http.json<IBankIdCollectApiResponse>(url, request))
-        .then(validate);
-};
+): Promise<IBankIdCollectApiResponse> =>
+    doCollectRequest(requestOptions).then(validate);
 
 export const cancel = (
     requestOptions: IBankIdCancelRequest
