+++ src/http/apis/bankid/collect.ts
@@ -1,0 +1,34 @@
+import * as http from "../../index";
+import Configuration from "../../../config";
+import {
+    IBankIdCollectRequest,
+    IBankIdCollectApiResponse,
+} from "../../../bankid/types";
+
+export const buildRequest = (): RequestInit => {
+    const requestForgeryToken = http.context().requestForgeryToken;
+    const request = http
+        .builder()
+        .method("post")
+        .accept("application/json")
+        .requestForgeryToken(requestForgeryToken)
+        .build();
+    return request;
+};
+
+export const getUrl = (orderRef: string) => {
+    const host = Configuration.current().getApiAddress();
+    const url = `${host}/bankid/collect/${orderRef}`;
+    return url;
+};
+
+export const collect = (
+    requestOptions: IBankIdCollectRequest
+): Promise<http.IApiResponse<IBankIdCollectApiResponse>> => {
+    const url = getUrl(requestOptions.orderRef);
+    const request = buildRequest();
+
+    return http.captureStateContext(
+        http.json<IBankIdCollectApiResponse>(url, request)
+    );
+};
